Split properties file
