+++ docs/components/Property.md
@@ -1,0 +1,11 @@
+Renders some property:
+
+```js
+<Property name="name" value="value" />
+```
+
+Name is optional:
+
+```js
+<Property value="value" />
+```
