+++ docs/components/PropertyName.md
@@ -1,0 +1,5 @@
+Renders name of a property.
+
+```js
+<PropertyName name="some-name" />
+```
